Adopt new version of dmon
